+++ inst/viewer/gene_expression/event_projection_update_plot.R
@@ -1,17 +1,8 @@
 ##----------------------------------------------------------------------------##
 ## Update projection plot when expression_projection_data_to_plot() changes.
 ##----------------------------------------------------------------------------##
-expression_projection_started <- reactiveVal(FALSE)
-observeEvent(
-  input[["expression_projection_render_request"]],
-  {
-    expression_projection_started(TRUE)
-  },
-  ignoreInit = TRUE
-)
-
 observe({
-  req(expression_projection_started())
+  req(input[["expression_projection_render_request"]])
   data <- expression_projection_data_to_plot()
   req(data)
   expression_projection_parameters_other[['reset_axes']] <- FALSE