+++ inst/viewer/spatial/event_projection_update_plot.R
@@ -2,17 +2,8 @@
 ## Update projection plot when spatial_projection_data_to_plot() changes.
 ##----------------------------------------------------------------------------##
 
-spatial_projection_started <- reactiveVal(FALSE)
-observeEvent(
-  input[["spatial_projection_render_request"]],
-  {
-    spatial_projection_started(TRUE)
-  },
-  ignoreInit = TRUE
-)
-
 observe({
-  req(spatial_projection_started())
+  req(input[["spatial_projection_render_request"]])
   data <- spatial_projection_data_to_plot()
   req(data)